combine quotebox and quoteboxcontainer components; add github link component
