+++ src/components/QuoteBox.jsx
@@ -1,9 +1,37 @@
+import { useState } from "react";
 import { Card, Stack, Button, Image, Fade } from "react-bootstrap";
 import styles from "../styles/QuoteBox.module.css";
+import quotes from "../../data/quotes.json";
 import quotationImg from "../assets/quotation.png";
 import githubLogo from "../assets/github-logo.png";
 
-function QuoteBox(props) {
+export default function QuoteBox() {
+  const getRandomQuoteIndex = () => Math.floor(Math.random() * quotes.length);
+
+  const [quoteIndex, setQuoteIndex] = useState(getRandomQuoteIndex());
+  const [textVisible, setTextVisible] = useState(true);
+
+  const handleNewQuoteButtonClick = () => {
+    const updateQuoteIndex = () => {
+      let newQuoteIndex = getRandomQuoteIndex();
+      while (newQuoteIndex === quoteIndex) {
+        newQuoteIndex = getRandomQuoteIndex();
+      }
+      setQuoteIndex(newQuoteIndex);
+    };
+
+    // fades text (quote and author) out
+    setTextVisible(false);
+    setTimeout(() => {
+      updateQuoteIndex();
+
+      // fades text back in
+      setTextVisible(true);
+    }, 550);
+  };
+
+  const selectedQuote = quotes[quoteIndex];
+
   return (
     <Stack gap={4} className={styles.stack}>
       <Card className={`p-2 ${styles.card}`}>
@@ -14,10 +42,10 @@ function QuoteBox(props) {
             className={styles.quotationImage}
           />
           <div className={styles.textContainer}>
-            <Fade in={props.textIsVisible}>
+            <Fade in={textVisible}>
               <Card.Text className={styles.cardText}>
-                <p>{props.text}</p>
-                <p className={styles.author}>- {props.author}</p>
+                <p>{selectedQuote.text}</p>
+                <p className={styles.author}>- {selectedQuote.author}</p>
               </Card.Text>
             </Fade>
           </div>
@@ -30,7 +58,7 @@ function QuoteBox(props) {
       </Card>
       <Button
         variant="primary"
-        onClick={props.handleNewQuoteButtonClick}
+        onClick={handleNewQuoteButtonClick}
         className={styles.button}
       >
         New Quote
@@ -42,4 +70,4 @@ function QuoteBox(props) {
   );
 }
 
-export default QuoteBox;
+// }
